Direct users to the official ammap package
Why:

I created this package because there wasn't an officially supported one
provided by amcharts. Happily, that is no longer the case and users
should instead use the official package

This PR:

Adds a note at the top of the readme to direct users to use the
officially supported ammap bower package.